backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1

# Vector Database
pinecone==5.4.1

# Embeddings
requests==2.32.5
tiktoken==0.8.0

# LLM
groq==0.11.0

# HTTP Client
httpx==0.27.2

# Utilities
pydantic==2.9.2
pydantic-settings==2.6.0
